+++ packages/kitchensink/service-fetchq-task/service-fetchq.test.js
@@ -1,0 +1,9 @@
+describe('service-fetchq-task', () => {
+  beforeEach(fetchq.resetState);
+
+  it('Should run a task', async () => {
+    console.log('ok');
+    // const res = await get.debug('/fetchq/append/q1/foobar');
+    // expect(res).toMatchObject({ subject: expect.any(String) });
+  });
+});