+++ packages/service-fetchq-task/src/on-init-service.js
@@ -6,11 +6,23 @@
  * @param {*} param0
  */
 module.exports = ({ createExtension, getConfig, setContext }) => {
+  const queueName = getConfig('fetchq.task.queue.name', 'task');
+  const queueSettings = getConfig('fetchq.task.queue.settings', {});
+  const workerSettings = getConfig('fetchq.task.worker.settings', {});
+
+  // Ensure a list of tasks is provided
+  let configuredTasks = getConfig('fetchq.task.register', []);
+  if (!Array.isArray(configuredTasks)) {
+    configuredTasks = [configuredTasks];
+  }
+
+  // TODO: check task correctness
+
   const registerTasks = [
     // Collect from App configuration
-    ...getConfig("fetchq.task.register", []),
+    ...configuredTasks,
     // Collect from other extensions
-    ...createExtension.sync("$FETCHQ_REGISTER_TASK").map(($) => $[0])
+    ...createExtension.sync('$FETCHQ_REGISTER_TASK').map(($) => $[0]),
   ];
 
   // TODO: validate tasks DTO
@@ -20,5 +32,10 @@ module.exports = ({ createExtension, getConfig, setContext }) => {
     throw new Error(`FetchqTask invalid format`);
   }
 
-  setContext("fetchq.task.register", registerTasks);
+  setContext('fetchq.task', {
+    queueName,
+    queueSettings,
+    workerSettings,
+    register: registerTasks,
+  });
 };